fix %lld when off_t is used on 64-bit (Alon Bar-Lev)

+++ src/ntfs-3g.c
@@ -703,7 +703,7 @@ static int ntfs_fuse_write(const char *org_path, const char *buf, size_t size,
 		res = ntfs_attr_pwrite(na, offset, size, buf);
 		if (res < (s64)size)
 			ntfs_log_perror("ntfs_attr_pwrite partial write (%lld: "
-				"%lld <> %d)", (s64)offset, (s64)size, res);
+				"%lld <> %d)", (long long)offset, (long long)size, res);
 		if (res <= 0) {
 			res = -errno;
 			goto exit;
